% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{falcon02_block}
\alias{falcon02_block}
\title{Falcon 02 block}
\usage{
falcon02_block(data, ...)
}
\arguments{
\item{data}{ADAM dataset.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Falcon 02 block
}
