% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{falcon05_block}
\alias{falcon05_block}
\title{Falcon 05 block}
\usage{
falcon05_block(data, ...)
}
\arguments{
\item{data}{ADAM dataset.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Falcon 05 block
}
