% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_02.R
\name{new_cardinal02_block}
\alias{new_cardinal02_block}
\title{FDA Table 2: Baseline Demographic and Clinical Characteristics, Safety Population, Pooled Analyses}
\usage{
new_cardinal02_block(
  ...,
  arm_var = character(),
  vars = character(),
  saffl_var = character(),
  lbl_overall = "Total Population"
)
}
\arguments{
\item{arm_var}{ARM.}

\item{vars}{Variables.}

\item{saffl_var}{Safety flag.}
}
\description{
FDA Table 2: Baseline Demographic and Clinical Characteristics, Safety Population, Pooled Analyses
}
