% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_03.R
\name{new_cardinal03_block}
\alias{new_cardinal03_block}
\title{FDA Table 3: Patient Screening and Enrollment, Trials A and B}
\usage{
new_cardinal03_block(
  ...,
  id_var = character(),
  arm_var = character(),
  scrnfl_var = character(),
  scrnfail_var = character(),
  scrnfailfl_var = character(),
  show_colcounts = FALSE,
  prune_0 = TRUE
)
}
\arguments{
\item{id_var}{ARM.}

\item{arm_var}{ARM.}

\item{scrnfl_var}{Screen flag.}

\item{scrnfailfl_var}{Screen fail flag.}

\item{show_colcounts}{Display counts.}

\item{vars}{Variables.}

\item{purne_0}{Prune zero rows.}
}
\description{
FDA Table 3: Patient Screening and Enrollment, Trials A and B
}
