% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_04.R
\name{new_cardinal04_block}
\alias{new_cardinal04_block}
\title{FDA Table 4: Patient Disposition, Pooled Analyses}
\usage{
new_cardinal04_block(
  ...,
  id_var = character(),
  arm_var = character(),
  pop_vars = character(),
  lbl_pop_vars = character(),
  show_colcounts = FALSE,
  prune_0 = TRUE
)
}
\arguments{
\item{id_var}{ARM.}

\item{arm_var}{ARM.}

\item{pop_vars}{Population variables.}

\item{lbl_pop_vars}{Population labels.}

\item{show_colcounts}{Display counts.}

\item{purne_0}{Prune zero rows.}
}
\description{
FDA Table 4: Patient Disposition, Pooled Analyses
}
