% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_05.R
\name{new_cardinal05_block}
\alias{new_cardinal05_block}
\title{FDA Table 5: Duration of Treatment Exposure, Safety Population, Pooled Analyses}
\usage{
new_cardinal05_block(
  show_colcounts = character(),
  arm_var = character(),
  id_var = character(),
  saffl_var = character(),
  trtsdtm_var = character(),
  trtedtm_var = character(),
  u_trtdur = "days",
  ...
)
}
\arguments{
\item{show_colcounts}{Display counts.}

\item{arm_var}{ARM.}

\item{id_var}{ARM.}

\item{vars}{Variables.}

\item{scrnfl_var}{Screen flag.}

\item{scrnfailfl_var}{Screen fail flag.}

\item{purne_0}{Prune zero rows.}
}
\description{
FDA Table 5: Duration of Treatment Exposure, Safety Population, Pooled Analyses
}
