% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_06.R
\name{new_cardinal06_block}
\alias{new_cardinal06_block}
\title{FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses}
\usage{
new_cardinal06_block(
  ...,
  id_var = character(),
  arm_var = character(),
  show_colcounts = FALSE,
  prune_0 = TRUE,
  saffl_var = character()
)
}
\description{
FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses
}
