% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_block.R
\name{new_label_block}
\alias{new_label_block}
\title{Label block constructor}
\usage{
new_label_block(value = NULL, ...)
}
\arguments{
\item{value}{Default value.}

\item{...}{Further (metadata) attributes}
}
\description{
Leverages the \link{keyvalue_field}
}
