% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{random_cdisc_data_block}
\alias{random_cdisc_data_block}
\title{Random CDISC data}
\usage{
random_cdisc_data_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Random CDISC data
}
