% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R, R/label_block.R
\name{relabel_block}
\alias{relabel_block}
\title{Relabel block}
\usage{
relabel_block(data, ...)

relabel_block(data, ...)
}
\arguments{
\item{data}{Dataset to relabel.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}.}
}
\description{
Relabel block

Relabel block
}
